+++ back/src/__tests__/mocks/ssoServer.ts
@@ -138,7 +138,6 @@ const handlers = [
       )
     }
   ),
-
   http.post('http://localhost:8000/api/v1/user', async ({ request }) => {
     const { authToken: token } = (await request.json()) as {
       authToken: string
@@ -172,6 +171,26 @@ const handlers = [
       } as GetUserResponse,
       { status: 200 }
     )
+  }),
+  http.get('http://localhost:8000/api/v1/itinerary', async () => {
+    const itineraries = [
+      {
+        id: 'clpb8tgmt000308l98pxn1rgj',
+        name: 'Backend Node.js',
+        slug: 'backend-nodejs',
+      },
+      {
+        id: 'clpb8tc17000208l90u4h0wqn',
+        name: 'Backend Java',
+        slug: 'backend-java',
+      },
+      {
+        id: 'clpb8t7kv000108l942sb5voy',
+        name: 'Frontend React',
+        slug: 'frontend-react',
+      },
+    ]
+    return HttpResponse.json(itineraries, { status: 200 })
   }),
 ]