+++ packages/ui/src/components/atoms/index.ts
@@ -2,5 +2,8 @@ import Input from './Input'
 import Label from './Label'
 import { Icon } from './Icon'
 import { BackButton } from './BackButton'
+import { Button } from './Button'
 
-export { Input, Label, Icon, BackButton }
+export { Input, Label, Button, Icon, BackButton }
+
+